requirements.txt
gradio>=4.0.0
openai>=1.0.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
PyPDF2>=3.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0